+++ source/annotatorlib/include/AnnotatorLib/ImageFolder.h
@@ -36,7 +36,7 @@ class ANNOTATORLIB_API ImageFolder : public ImageSet {
    * @param frame
    * @return image
    */
-  virtual Image getImage(unsigned int /*in*/ frame) override;
+  virtual Image getImage(unsigned long /*in*/ frame) override;
 
   virtual bool gotoPosition(unsigned long position) override;
